Code added to GCode_Interpreter to support stepper-driven extruders.  The ordinary DC motor speed (pulse = step) and direction (= direction) pins also drive the stepper.  Ther is an additional stepper-enable pin defined.  Set this -ve (the default) to not use a stepper drive.  Set it to the enable pin for the stepper to use one.

+++ trunk/reprap/firmware/GCode_Interpreter/process_string.pde
@@ -63,7 +63,7 @@ byte x_direction = 1;
 byte y_direction = 1;
 byte z_direction = 1;
 
-int extruder_speed = 0;
+float extruder_speed = 0;
 
 int scan_int(char *str, int *valp);
 int scan_float(char *str, float *valp);
@@ -557,7 +557,7 @@ void process_string(char instruction[], int size)
 				//set max extruder speed, 0-255 PWM
 			case 108:
 				if (gc.seen & GCODE_S)
-					extruder_speed = (int)gc.S;
+					extruder_speed = gc.S;
 				break;
 
                                 // Open the valve
